refactor: 타입 변경으로 인한 리팩토링
