+++ user-api/src/main/java/org/ecommerce/userapi/dto/SellerDto.java
@@ -2,7 +2,7 @@ package org.ecommerce.userapi.dto;
 
 import java.time.LocalDateTime;
 
-import org.ecommerce.userapi.entity.type.UserStatus;
+import org.ecommerce.userapi.entity.enumerated.UserStatus;
 import org.ecommerce.userapi.exception.UserErrorMessages;
 import org.ecommerce.userapi.security.JwtUtils;
